rename module ui -> tui add Builder for TuiEngine use u16 instead of u32 in terminal

+++ tui/src/attr.rs
@@ -1,8 +1,8 @@
 #[derive(Clone, Debug, PartialEq, PartialOrd, Ord, Eq, Hash)]
 pub struct Attr {
     pub id: String,
-    pub width: u32,
-    pub height: u32,
+    pub width: u16,
+    pub height: u16,
     pub display: AttrDisplay,
     pub float: bool,
 }
